Error handling added

+++ master-data/src/main/java/org/weekendsoft/mpa/masterdata/controller/AccountController.java
@@ -7,12 +7,18 @@ import java.util.List;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.weekendsoft.mpa.masterdata.exception.ERROR_CODES;
+import org.weekendsoft.mpa.masterdata.exception.RecordNotFoundException;
 import org.weekendsoft.mpa.masterdata.model.Account;
+import org.weekendsoft.mpa.masterdata.model.ErrorInfo;
 import org.weekendsoft.mpa.masterdata.repository.AccountRepository;
 
 /**
@@ -25,6 +31,16 @@ public class AccountController {
 	
 	@Autowired
 	private AccountRepository accountRepository;
+	
+	@ResponseStatus(HttpStatus.NOT_FOUND)
+	@ExceptionHandler(RecordNotFoundException.class)
+	public ErrorInfo handleException(RecordNotFoundException ex) {
+        ErrorInfo error = new ErrorInfo();
+        error.setCode(ERROR_CODES.RECORD_NOT_FOUND);
+        error.setMessage("Record with id: " + ex.id);
+        error.setStatus(HttpStatus.NOT_FOUND);
+        return error;
+    }
 	
 	@RequestMapping("")
 	public String home() {
@@ -43,12 +59,17 @@ public class AccountController {
 	
 	@RequestMapping(value = "accounts/{id}", method = RequestMethod.GET)
 	public Account get(@PathVariable int id) {
-		return accountRepository.findOne(id);
+		Account account = accountRepository.findOne(id);
+		if (account == null) throw new RecordNotFoundException(id, "Account ID not found: " + id, null);
+		
+		return account;
 	}
 	
 	@RequestMapping(value = "accounts/{id}", method = RequestMethod.PUT)
 	public Account update(@PathVariable int id, @RequestBody Account account) {
 		Account existingAccount = accountRepository.findOne(id);
+		if (existingAccount == null) throw new RecordNotFoundException(id, "Account ID not found: " + id, null);
+		
 		BeanUtils.copyProperties(account, existingAccount);
 		return accountRepository.saveAndFlush(existingAccount);
 	}
@@ -56,6 +77,8 @@ public class AccountController {
 	@RequestMapping(value = "accounts/{id}", method = RequestMethod.DELETE)
 	public Account delete(@PathVariable int id) {
 		Account existingAccount = accountRepository.findOne(id);
+		if (existingAccount == null) throw new RecordNotFoundException(id, "Account ID not found: " + id, null);
+		
 		accountRepository.delete(existingAccount);
 		return existingAccount;
 	}
